% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePval.R
\name{calculatePvalue}
\alias{calculatePvalue}
\title{Calculates P-value for the regions}
\usage{
calculatePvalue(trueTstat, compare_tstats)
}
\arguments{
\item{trueTstat}{The T-statistics value calculated using 
\code{\link{calculateTvalue}} for the intended comparison.}

\item{compare_tstats}{The T-statistics value calculated using 
\code{\link{calculateTvalue}} for all other comparisons possible.}
}
\value{
A vector that is the adjusted P-value for the intended comparison.
}
\description{
This calculates the adjusted P-values for the regions using column 
permutation and Benjamini Hochberg correction methods.
}
\seealso{
\code{\link{DBAnalyze}} which calls this function
}
