% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComparisons.R
\name{getComparisons}
\alias{getComparisons}
\title{Identify possible combinations}
\usage{
getComparisons(trueLabel, comparison, numSamples)
}
\arguments{
\item{trueLabel}{The true labels for the samples.}

\item{comparison}{A vector containing the comparison to be made. Names here 
need to correspond to the sample groups in the sample file (Eg. c("G1",G2") 
means the comparison G1/G2).}

\item{numSamples}{Number of samples in the dataset.}
}
\value{
A data frame with possible combinations of samples other the true 
intended comparison.
}
\description{
This function creates a data frame of all possbile combinations of sample 
labels.
This information is utilized by \code{\link{calculatePvalue}} to calculate 
the P-value using column permutation method.
}
\seealso{
\code{\link{DBAnalyze}} which calls this function and 
\code{\link{getNewLabels}} which this function calls
}
