% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewLabels.R
\name{getNewLabels}
\alias{getNewLabels}
\title{Labels the samples according to the combinations from \code{\link{getComparisons}}}
\usage{
getNewLabels(trueLabel, comparison, numSamples, combns, index)
}
\arguments{
\item{trueLabel}{The true labels for the samples.}

\item{comparison}{A vector containing the comparison to be made. Names here 
need to correspond to the sample groups in the sample file (Eg. c("G1",G2") 
means the comparison G1/G2).}

\item{numSamples}{Number of samples in the dataset.}

\item{combns}{Possible combinations of sample index generated in 
\code{\link{getComparisons}}.}

\item{index}{index of the combination to use for labeling.}
}
\value{
A vector with labels.
}
\description{
This function labels the samples according the combinations generated by 
\code{\link{getComparisons}}.
}
\seealso{
\code{\link{getComparisons}} which calls this function
}
