% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDist.R
\name{plotDist}
\alias{plotDist}
\title{Plot data distribution histograms}
\usage{
plotDist(countVec, ansVec, sampleName)
}
\arguments{
\item{countVec}{Vector containing the count data for a sample.}

\item{ansVec}{Vector containing the anscombe transformed count data for a 
sample.}

\item{sampleName}{Name of the sample being plotted.}
}
\value{
None
}
\description{
This function is to plot data distribution histogram before
and after anscombe transformation.
}
